Using inputChange Handler for all inputs
this will be used to handle the input types in the form and update the state of the component with data from the form. This will pipe all
the form state to the react component state and be able to keep the component aware of the internal state of the form. Note that this will
also be usefull when updating the state of the form inputs, as they are being updated from the internal state of the component.
Using 1 handler function to handle input changes, ensures that the application component is clearer and more terse and also allows for using
ES6 Computed Property Name to handle the changes of the component state.